requirements.txt
django
whitenoise
google-genai
groq
python-dotenv
gunicorn
django-cors-headers
django-allauth
PyJWT
cryptography
stripe
psycopg2-binary
dj-database-url
reportlab
